remove the unnecessary util file
